+++ mcp_tools/cloud_container/cloud_and_container_security.py
@@ -5,73 +5,6 @@ from typing import Dict, Any
 def register_cloud_and_container_security_tools(mcp, hexstrike_client, logger):
 
     @mcp.tool()
-    def kube_hunter_scan(target: str = "", remote: str = "", cidr: str = "",
-                        interface: str = "", active: bool = False, report: str = "json",
-                        additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute kube-hunter for Kubernetes penetration testing.
-
-        Args:
-            target: Specific target to scan
-            remote: Remote target to scan
-            cidr: CIDR range to scan
-            interface: Network interface to scan
-            active: Enable active hunting (potentially harmful)
-            report: Report format (json, yaml)
-            additional_args: Additional kube-hunter arguments
-
-        Returns:
-            Kubernetes penetration testing results
-        """
-        data = {
-            "target": target,
-            "remote": remote,
-            "cidr": cidr,
-            "interface": interface,
-            "active": active,
-            "report": report,
-            "additional_args": additional_args
-        }
-        logger.info(f"☁️  Starting kube-hunter Kubernetes scan")
-        result = hexstrike_client.safe_post("api/tools/kube-hunter", data)
-        if result.get("success"):
-            logger.info(f"✅ kube-hunter scan completed")
-        else:
-            logger.error(f"❌ kube-hunter scan failed")
-        return result
-
-    @mcp.tool()
-    def kube_bench_cis(targets: str = "", version: str = "", config_dir: str = "",
-                      output_format: str = "json", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute kube-bench for CIS Kubernetes benchmark checks.
-
-        Args:
-            targets: Targets to check (master, node, etcd, policies)
-            version: Kubernetes version
-            config_dir: Configuration directory
-            output_format: Output format (json, yaml)
-            additional_args: Additional kube-bench arguments
-
-        Returns:
-            CIS Kubernetes benchmark results
-        """
-        data = {
-            "targets": targets,
-            "version": version,
-            "config_dir": config_dir,
-            "output_format": output_format,
-            "additional_args": additional_args
-        }
-        logger.info(f"☁️  Starting kube-bench CIS benchmark")
-        result = hexstrike_client.safe_post("api/tools/kube-bench", data)
-        if result.get("success"):
-            logger.info(f"✅ kube-bench benchmark completed")
-        else:
-            logger.error(f"❌ kube-bench benchmark failed")
-        return result
-
-    @mcp.tool()
     def docker_bench_security_scan(checks: str = "", exclude: str = "",
                                   output_file: str = "/tmp/docker-bench-results.json",
                                   additional_args: str = "") -> Dict[str, Any]:
