+++ mcp_tools/k8s_scan/kube_hunter.py
@@ -1,0 +1,41 @@
+# mcp_tools/k8s_scan/kube_hunter.py
+
+from typing import Dict, Any
+
+def register_kube_hunter_tool(mcp, hexstrike_client, logger):
+    
+    @mcp.tool()
+    def kube_hunter_scan(target: str = "", remote: str = "", cidr: str = "",
+                        interface: str = "", active: bool = False, report: str = "json",
+                        additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute kube-hunter for Kubernetes penetration testing.
+
+        Args:
+            target: Specific target to scan
+            remote: Remote target to scan
+            cidr: CIDR range to scan
+            interface: Network interface to scan
+            active: Enable active hunting (potentially harmful)
+            report: Report format (json, yaml)
+            additional_args: Additional kube-hunter arguments
+
+        Returns:
+            Kubernetes penetration testing results
+        """
+        data = {
+            "target": target,
+            "remote": remote,
+            "cidr": cidr,
+            "interface": interface,
+            "active": active,
+            "report": report,
+            "additional_args": additional_args
+        }
+        logger.info(f"☁️  Starting kube-hunter Kubernetes scan")
+        result = hexstrike_client.safe_post("api/tools/kube-hunter", data)
+        if result.get("success"):
+            logger.info(f"✅ kube-hunter scan completed")
+        else:
+            logger.error(f"❌ kube-hunter scan failed")
+        return result
